Deprecate and move exported internal functions
These functions should not have been exported, but we must maintain the compatibility promise until v2.
